Fixed a number of stability/tolerance issues (hopefully.)
Note: The search algorithm in DriftShell.c (that tries to make I-I0 go to
zero), is (once again) a minimizer rather than a simple root finder.  The
reason for this is that it is hard to bracket the root for pitch angles near
zero. So rather than waste all that time, I'm using a minimizer again --
because its easier to bracket the minimum than it is to bracket the zero.
Also, Ive modified the tolerance in the Quality settings. Be careful with
these! The searching will fall down if you try to get do too high of a
tolerance on I-I0 with too low of a tolerance on the individiual I's.  Also, I
added a bit of a stochastic search capability (nothing too fancy) to help deal
with potentially pathological cases. An finally, I added a limit on total
number of iterations in the search to avoid infinite loops when things go crazy
for whatever reason.