Fairly major updates to Lstar calculations.
1) Added an empirical fitting procedure to try and accelerate root-finding.  Used
LS type fitting to a quadtratic in I-I0 versus mlat. Then try to jump straight
to mlat that gives I-I0=0. This is similar to brents method for root solving,
but here we dont want the quadtraic curve to go through each point (each of
which may have significant uncertainties).

2) Also changed rational function extrap to polynomial function extrap in
bulirsch stoer method. Seesm that rat func ext. can fail catastrophically at
times (rare, but can happen). On the other hand, poly extra doesnt seem to be
as accurate. We should probably try both and pich the one with smallest error
estimate (which is kept track of).

3) Still need to figure out how aggressively we need to use the fitting in the
bisection/minimization phases.

4) Still need to do some testing.