Fixed a bug in TraceLine.c
When the initial point is already below your target stopping height,
traceline tries to climb above it first. But it was trying with a fixed
Htry of 0.1. This causes problems in (rare?) cases where you are below
to start and the length of the FL between foorpoints is less than 0.1
Re. I added a counter to bail if too many iterations, but also added
code to detect when the initial difference in height is lower than 0.1.
In such cases it tries a step of 0.5 of the difference to get above the
target height.