TS04 tracing issues fixed(???)
TS04 seems to get into weird field regions near the magnetopause(?) We end up
tracing forever it seems. Put a limit on how long the field line can be before
we call it open. This seems to dfix things for now...

In future we probably want to have a seperate flag for those types of FLS.