Bug fix in Lgm_MagEphemWriteHdf.c and added option to MagEphemFromSpiceKernel.c
There was a bug when writing headers in Lgm_MagEphemWriteHdf.c. The number of times in a few variables were not correct.

Also, added "-T" option to MagEphemFromSpiceKernel.c. This allows user to fix the field model at a given time. Usefull for comparing differences between static and dynamic versions of the same model.