Changed search algoirthm in FindBmRadius().
The initial algorthim did an intial bracketing then a bisection to find the location where B==Bm.
In the outer regions of the magnetosphere, the initial bracketing could sometimes skip over the
root. To fix this, I added the ability to track local minima as well. Now, if a
zero bracket is not initially detected, the algorithm search the local minima
that were encounetered. Seems ,to fix some stabilityl issues for LCDS calcs, etc...