Modifications to Lgm_SummersDiffCoeff.c to improve the efficiency of the computation of the local and bounce-averaged Glauert and Horne diffusion coefficients.
Several routines have been modified so that an input parameter to the routine is the plasma parameter, aStar=(wce/wpe)^2, in place of sending in both the cyclotron frequency, wce, and plasma frequency, wpe, separately.  The electronColdPlasmaGroupVelocity subroutine has been modified in this way.  The normalizer, N(w), given by equation (15) of Glauert and Horne's 2005 JGR paper, is now computed in this way by the routine normalizerForWavePowerSpectrumParams, but the result from this subroutine must by multiplied by wce^2 in order to reproduce equation (15), and so the calling subroutines have also been modified to multiply the new N(w) by wce^2. The normalizer is now tabulated on a grid in normalized frequency, w/wce, and aStar, i.e. N=N(w/wce,aStar), and this table of values is used instead of computing the normalizer at each point in the integration with respect to latitude and wave normal angle. The parameters that specify the grid in normalized frequency and aStar are now part of the Lgm_SummersInfo C struct defined in Lgm_SummersDiffCoeff.h (the new parameters are called aStarMin, aStarMax, nPlasmaParameters, nNw, and a pointer to the Nw table, *Nw), and these parameters are passed as arguments to the high level routine Lgm_GlauertandHorneDxxBounceAvg and Lgm_GlauertAndHorneHighFrequencyDiffusionCoefficients_Local so that the table can be properly interpreted. Note that the table should always uniformly sample the normalized frequency range [0,1] even though the routine that is used to compute the table allows the user to specify alternate ranges in normalized frequency. This is because the routines localDiffusionCoefficientAtSpecificThetaGlauertAndHorne and localDiffusionCoefficientAtSpecificThetaGlauertAndHorneReturnTensorInPlace assume that the range in normalized frequency is [0,1] when computing an index into the array in order to evaluate N(w). Also note that the sampling in aStar is assumed to be geometrical.

The approach documented in Appendix A of Jay Albert's 2007 JGR paper 'Simple approximations of quasi-linear diffusion coefficients', wherein the range of wave normal angles in which a resonant root can exist given a certain frequency range, is now correctly implemented in the routine computeWaveNormalAnglesWhereResonantRootsMayExist, and used in the routine Lgm_GlauertAndHorneHighFrequencyDiffusionCoefficients_Local. Since this approach can only be applied for a single harmonic number, the code has been re-written to peform the integral over wave normal angle using the derived range for each harmonic number rather than performing the integral over wave normal angle using the sum over all harmonic numbers, as was done previously. The method only reduces the total computation time by a small amount (10's of percent).

The routines localDiffusionCoefficientAtSpecificThetaGlauertAndHorne and localDiffusionCoefficientAtSpecificThetaGlauertAndHorneReturnTensorInPlace now have tests for the local pitch angle being too close to 0 or 90 degrees, which can cause divide by zero errors, or the wave normal angle being too close to 0 degrees, which causes the polynomial for which we want to find the roots to be degenerate. In any of these cases, the angle is set to be 1e-3 different than the problematic case (in radians).

The routine computeResonantRootsUsingColdElectronPlasma has modified the polynomial for which the roots are desired by multiplying the old polynomial by a large power of the cosine of the wave normal angle so that there is no longer a cosine of the wave normal angle in the denominator of any coefficient. The Landau resonance (n=0) is now treated specially since the polynomial is degenerate in this case and a lower-order polynomial is adequate.